+++ test/components/alpha/button_marketing_test.rb
@@ -2,41 +2,41 @@
 
 require "test_helper"
 
-class PrimerAlphaButtonMarketingComponentTest < Minitest::Test
+class PrimerAlphaButtonMarketingTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_renders_content
-    render_inline(Primer::Alpha::ButtonMarketingComponent.new) { "content" }
+    render_inline(Primer::Alpha::ButtonMarketing.new) { "content" }
 
     assert_selector("button.btn-mktg", text: "content")
   end
 
   def test_defaults_button_tag_with_scheme
-    render_inline(Primer::Alpha::ButtonMarketingComponent.new) { "content" }
+    render_inline(Primer::Alpha::ButtonMarketing.new) { "content" }
 
     assert_selector("button.btn-mktg[type='button']")
   end
 
   def test_renders_a_as_a_button
-    render_inline(Primer::Alpha::ButtonMarketingComponent.new(tag: :a)) { "content" }
+    render_inline(Primer::Alpha::ButtonMarketing.new(tag: :a)) { "content" }
 
     assert_selector("a.btn-mktg[role='button']")
   end
 
   def test_renders_href
-    render_inline(Primer::Alpha::ButtonMarketingComponent.new(href: "www.example.com")) { "content" }
+    render_inline(Primer::Alpha::ButtonMarketing.new(href: "www.example.com")) { "content" }
 
     assert_selector("button[href='www.example.com']")
   end
 
   def test_renders_with_the_css_class_mapping_to_the_provided_type
-    render_inline(Primer::Alpha::ButtonMarketingComponent.new(scheme: :primary)) { "content" }
+    render_inline(Primer::Alpha::ButtonMarketing.new(scheme: :primary)) { "content" }
 
     assert_selector(".btn-mktg.btn-primary-mktg")
   end
 
   def test_renders_with_the_css_class_variant_mapping_to_the_provided_variant
-    render_inline(Primer::Alpha::ButtonMarketingComponent.new(variant: :large)) { "content" }
+    render_inline(Primer::Alpha::ButtonMarketing.new(variant: :large)) { "content" }
 
     assert_selector(".btn-mktg.btn-large-mktg")
   end